newsimages25/workflows/DACS-UM-RTL_RETRIEVAL/requirements.txt
accelerate==1.3.0
bitsandbytes==0.45.2
datasets==3.3.2
faiss-cpu==1.11.0
pandas==2.2.3
numpy==1.25.2
peft==0.14.0
pillow==11.1.0
pytest==8.3.4
seaborn==0.13.2
matplotlib==3.10.0
num2words==0.5.14
sentencepiece==0.2.0
scipy==1.15.1
scikit-learn==1.2.2
tqdm==4.67.1
protobuf==6.30.2
pytorch-lightning==2.5.0.post0
torch==2.3
torchmetrics==1.6.1
torchvision==0.18
transformers==4.50.0
ultralytics==8.3.76
wandb==0.20.1
einops==0.8.1
open_clip_torch==3.1.0
